+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## unreleased
+### Added
+* Added support for xbc1 compressed `BC` animations to `xc3_model::load_animations`. 
+
 ### Changed
 * Changed skeleton creation to only produce a single root bone for improved compatibility.